% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{Draw.Panels}
\alias{Draw.Panels}
\title{Draw 4-panel graph}
\usage{
Draw.Panels(date, flow, rain, H = NA)
}
\arguments{
\item{date}{vector of date for flow and rain data}

\item{flow}{vector of flow data, in GPD}

\item{rain}{vector of rain, in inches}

\item{H}{unit hydrograph from output (optional)}
}
\value{
unit hydrograph for each event
}
\description{
Draw 4-panel graph
}
\examples{
data(DF)
\dontrun{Draw.Panels(DF$date, DF$Hollybrook, DF$rain)}
}
