% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSOAP.R
\name{Get.Cost}
\alias{Get.Cost}
\title{Cost Function for GA optimizer}
\usage{
Get.Cost(Q, PU, DateRange, x1, x2, x3, x4, x5, x6, x7, x8, x9, x10, x11,
  x12, x13, x14, x15)
}
\arguments{
\item{Q}{Daily flow, in GPD}

\item{PU}{Square rainfall matrix}

\item{DateRange}{Time Window for analysis}

\item{x1}{Day 01}

\item{x2}{Day 02}

\item{x3}{Day 03}

\item{x4}{...}

\item{x5}{...}

\item{x6}{...}

\item{x7}{...}

\item{x8}{...}

\item{x9}{...}

\item{x10}{...}

\item{x11}{...}

\item{x12}{...}

\item{x13}{Day 13}

\item{x14}{Day 14}

\item{x15}{Day 15}
}
\value{
Mean-Squared Error between actual flow and hydrograph flow
}
\description{
Cost Function for GA optimizer
}
