% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSOAP.R
\name{Get.Events}
\alias{Get.Events}
\title{Identify major RDII events}
\usage{
Get.Events(date, flow, rain)
}
\arguments{
\item{date}{vector of date for flow and rain data}

\item{flow}{vector of flow data, in GPD}

\item{rain}{vector of rain, in inches}
}
\value{
vector of significant, single day events
}
\description{
Identify major RDII events
}
\examples{
data(DF)
Get.Events(DF$date, DF$Buffalo, DF$rain)
}
