% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSOAP.R
\name{Get.Hydrograph}
\alias{Get.Hydrograph}
\title{Get Hydrograph}
\usage{
Get.Hydrograph(Q, PU, DateRange)
}
\arguments{
\item{Q}{Daily flow, in GPD}

\item{PU}{Square rainfall matrix}

\item{DateRange}{Time Window for analysis}
}
\value{
15-day Unit Hydrograph
}
\description{
Get Hydrograph
}
\examples{
data(DF)
Ev <- Get.Events(DF$date, DF$Barnes, DF$rain)
RD <- Get.RDII(DF$date, DF$Barnes, DF$rain)
PU <- Lag.Rain(DF$rain)
\dontrun{Get.Hydrograph(RD, PU, -21:21+Ev[1])}
}
