% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSOAP.R
\name{Get.Rain}
\alias{Get.Rain}
\title{Get rainfall matrix}
\usage{
Get.Rain(P)
}
\arguments{
\item{P}{Daily Rainfall, in inches}
}
\value{
A square matrix of lagged rainfall hyetographs
}
\description{
Get rainfall matrix
}
\examples{
data(DF)
Get.Rain(DF$rain)
}
