% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{Get.Summary}
\alias{Get.Summary}
\title{Summarize Wet-Weather Flow}
\usage{
Get.Summary(date, flow, rain, H = NA)
}
\arguments{
\item{date}{vector of date for flow and rain data}

\item{flow}{vector of flow data, in GPD}

\item{rain}{vector of rain, in inches}

\item{H}{unit hydrograph from output (optional)}
}
\value{
list
}
\description{
Summarize Wet-Weather Flow
}
\examples{
data(DF)
\dontrun{Get.Summary(DF$date, DF$Hollybrook, DF$rain)}
}
