% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSOAP.R
\name{induce_daily_dwf}
\alias{induce_daily_dwf}
\title{Calculates Dry-Weather Flow (DWF)}
\usage{
induce_daily_dwf(date, flow, rain, max.rain.short = 0,
  dry.days.short = 7, max.rain.long = 1, dry.days.long = 14,
  max.stdev = 0.5)
}
\arguments{
\item{date}{vector of date for flow and rain data}

\item{flow}{vector of flow data, in GPD}

\item{rain}{vector of rain, in inches}

\item{max.rain.short}{maximum allowable rain during the short-term}

\item{dry.days.short}{number of days considered in the short-term}

\item{max.rain.long}{maximum allowable rain during the long-term}

\item{dry.days.long}{number of days considered in the long-term}

\item{max.stdev}{maximum allowable standard deviations in any term}
}
\value{
a dataframe of DWF by weekend and weekday
}
\description{
Calculates Dry-Weather Flow (DWF)
}
\examples{
data(DF)
induce_daily_dwf(DF$date, DF$Buffalo, DF$rain)
}
