% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSOAP.R
\name{infer_daily_gwi}
\alias{infer_daily_gwi}
\title{Calculate Groundwater Intrusion (GWI)}
\usage{
infer_daily_gwi(date, flow, rain)
}
\arguments{
\item{date}{vector of date for flow and rain data}

\item{flow}{vector of flow data, in GPD}

\item{rain}{vector of rain, in inches}
}
\value{
vector of flow minus DWF
}
\description{
Calculate Groundwater Intrusion (GWI)
}
\examples{
data(DF)
infer_daily_gwi(DF$date, DF$Buffalo, DF$rain)
}
