% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSOAP.R
\name{infer_daily_rdii}
\alias{infer_daily_rdii}
\title{Get Rainfall-Derived Inflow and Infiltration (RDII)}
\usage{
infer_daily_rdii(date, flow, rain)
}
\arguments{
\item{date}{vector of date for flow and rain data}

\item{flow}{vector of flow data, in GPD}

\item{rain}{vector of rain, in inches}
}
\value{
vector of flow minus DWF minus GWI
}
\description{
Get Rainfall-Derived Inflow and Infiltration (RDII)
}
\examples{
data(DF)
infer_daily_rdii(DF$date, DF$Buffalo, DF$rain)
}
