% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDII.R
\name{RTK}
\alias{RTK}
\title{Generates a Triangular Unit Hydrograph.}
\usage{
RTK(r, t, k)
}
\arguments{
\item{r}{Rainfall Volume, in percent}

\item{t}{Time to Peak, in days}

\item{k}{Recession Constant}
}
\value{
Unit Hydrograph
}
\description{
Generates a Triangular Unit Hydrograph.
}
\examples{
plot(RTK(0.025,2,4), type="l")
}
