% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Draw.R
\name{Draw.Panels}
\alias{Draw.Panels}
\title{Draw 4-panel graph}
\usage{
Draw.Panels(Q, H = NA)
}
\arguments{
\item{Q}{daily flow, in GPD}

\item{H}{unit hydrograph from output (optional)}
}
\value{
unit hydrograph for each event
}
\description{
Draw 4-panel graph
}
\examples{
data(DF)
H <- Draw.Panels(DF$Hollybrook)
}
